мелкие доработки create.py

+++ commands/create.py
@@ -1,10 +1,13 @@
 import os
 
+
 def create(file_path):
-    try:
-        with open(file_path, "w") as f:
-            f.write('')
-    except FileExistsError:
-        os.makedirs(os.path.dirname(file_path), exist_ok=True)
-        with open(file_path, "w") as f:
-            f.write(f'Файла {file_path} не существует, но мы создали новый')
+    if os.path.exists(file_path):
+        print('Such a file already exists')
+    else:
+        try:
+            with open(file_path, "w") as f:
+                f.write('')
+                print('Your file has been created')
+        except FileExistsError:
+            print('An error occurred while creating the file')
